shore: cleanup cmd & move cmd in its own package
