+++ cmd/cloudstic/store_test.go
